+++ Test_files/main.h
@@ -5,8 +5,8 @@ int _printf(const char *format, ...);
 int _putchar(char c);
 int print_string(char *);
 int print_decimal(int value);
-long long int power(int base, int x);
-long long convert_to_decimal(int Number, int base);
+long int power(int base, int x);
+long convert_to_decimal(int Number, int base);
 int print_binary(int number);
 int print_unsigned(unsigned int number);
